Fixed Location class and added test cases.
Changed Location to store its values as NSNumbers and NOT scalars (doubles). Wrote test cases up for it to make sure it calculates the other values properly. Spent a bit of time because got snagged up in minor bug. Used formatted string value %@ instead of  %f when trying to log doubles :(. But found it eventually and now it passes test!