Added stubs and TODOs to remind me to do them.
Basically Actions and Flags and Highlights all act as association objects between two entities. So want to add methods that connect them in the way that I want to use them. For example, I want to have a method that lets a Highlight act as a bridge between a Person and Place, and also has extra information attached to it like text and photos.