Setting up a standard way of using switch statements with values from enums.
Set a default gracefully handled value at the top. Go through and see if it fits the values I expect. If it doesn't, then NSAssert an error, which causes the program to crash in debug mode. But in the future, I can turn that off in release mode and instead it will gracefully handle it by the default value. In this case specifically the number of rows will default to 0, and a cell in a section that doesn't exist will return nil.